(refactor) Rename getCnnFeed to pullCnnFeed.
